GUI actualizado

+++ src/lab5_josuerocha/Ingenierias.java
@@ -10,7 +10,7 @@ public class Ingenierias extends Carrera{
     public Ingenierias() {
     }
 
-    public Ingenierias(String nombre, String facultad, double mensualidad, String jefe) {
+    public Ingenierias(String nombre, String facultad, int mensualidad, String jefe) {
         super(nombre, facultad, mensualidad, jefe);
     }
