+++ src/lab5_josuerocha/Salud.java
@@ -10,7 +10,7 @@ public class Salud extends Carrera{
     public Salud() {
     }
 
-    public Salud(String nombre, String facultad, double mensualidad, String jefe) {
+    public Salud(String nombre, String facultad, int mensualidad, String jefe) {
         super(nombre, facultad, mensualidad, jefe);
     }
